backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database — MongoDB async driver
motor==3.3.2
pymongo==4.6.1

# Redis — caching, sessions, rate limiting
redis[asyncio]==5.0.1
hiredis==2.3.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Razorpay SDK
razorpay==1.4.1

# Utilities
python-dotenv==1.0.0
bson==0.5.10

# Date/Time
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development / Linting
black==24.1.0
isort==5.13.2
flake8==7.0.0
